Misc (re #1559): fixed compile errors/warnings on Symbian platform.
git-svn-id: https://svn.pjsip.org/repos/pjproject/trunk@4411 74dad513-b988-da41-8d7b-12977e46ad98

+++ pjmedia/src/pjmedia/types.c
@@ -37,10 +37,10 @@ PJ_DEF(const char*) pjmedia_type_name(pjmedia_type t)
 	"unknown"
     };
 
-    pj_assert(t < PJ_ARRAY_SIZE(type_names));
+    pj_assert(t < (int)PJ_ARRAY_SIZE(type_names));
     pj_assert(PJMEDIA_TYPE_UNKNOWN == 4);
 
-    if (t < PJ_ARRAY_SIZE(type_names))
+    if (t < (int)PJ_ARRAY_SIZE(type_names))
 	return type_names[t];
     else
 	return "??";
